refactor: 프로젝트 불러오기 관련 파일 분리함
